pyproject.toml
[project]
name = "ntl-systoolbox"
version = "1.0.0"
description = "Outil CLI d'exploitation système pour Nord Transit Logistics"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Nord Transit Logistics", email = "it@nordtransit.local"}
]
keywords = ["cli", "sysadmin", "backup", "audit", "monitoring"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: System :: Monitoring",
    "Topic :: Database :: Database Engines/Servers",
]

dependencies = [
    "pyyaml>=6.0",
    "mysql-connector-python>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]
monitoring = [
    "psutil>=5.9",
]
all = [
    "ntl-systoolbox[dev,monitoring]",
]

[project.scripts]
ntl-systoolbox = "ntl_systoolbox.cli.__main__:main"
ntl-diag = "ntl_systoolbox.cli.__main__:main"

[project.urls]
Homepage = "https://github.com/nordtransit/ntl-systoolbox"
Documentation = "https://github.com/nordtransit/ntl-systoolbox#readme"
Repository = "https://github.com/nordtransit/ntl-systoolbox.git"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["ntl_systoolbox*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.venv
    | venv
    | build
    | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.flake8]
max-line-length = 100
exclude = [".git", "__pycache__", "build", "dist", "venv"]
ignore = ["E501", "W503"]
